Created the middle connection table that was on our diagram.
Connection table in question being QuestionsInQuiz.
TODO added for when deletion occurs and if question already exists. If
it exists, offer the user to use it? Regarding deletion, currently it
deletes the question from everybody. We need to be able to identify who
has made what question to be able to determine if the person in question
is authorized to delete the whole question.

There is still the now redundant connection between Question and Quiz
(there is quizId within the Question model).