Fetching of the currently being asked question has been implemented.
Within the Quiz the set was changed to a list which in the process has
made it so that there can be double entries of the same question (if the
host decides to make it so). Can easily be fixed with a check whenever a
question is being added. The check has not been implemented in case we
revert back to sets. Reason for going to list is for the order of
elements. If it goes back to a set how're we going to know the order of
elements when there will be no identifier on the questions themselves
that indicates number what they are in the quiz.

Incrementation of the currently being asked question has been
implemented.

So now when the users phone sends the request he'll only be able to get
the currently being asked question which is controlled by the host. Only
increments upwards by one and does not go out of bounds. (Stops on the
final question, debatable whether we want to add a flag or something
that indicates that the end has been reached or whether we have the host
end the quiz manually).