Registration and relevant checks have been implemented.
Checks whether a team with given name already exists for the quiz in
question. (Teams are allowed to have same name as long as it's not
within the same quiz). No need to check if phoneId already exists due to
offers of rejoining a quiz if you're registered for a quiz and if you
refuse you'd be removed from the quiz and therefore the phoneId would
not exist within a given quiz that the user is trying to register for.

Team registration now also saves the teams into the database.

Discussion needs to take place regarding the return of a JSON object for
when the team is registered.