Check whether the fetch request for the question is valid.
Checks whether the phone that is requesting a question from given quiz
is allowed to receive the question or not.

FetchQuestionWrapper still contains question number which is obsolete,
it has not been removed due to the fact that it'd break the client side
(due to them relying on sending x format JSON string). Also due to the
fact that the quizId may or may not be moved over to the URL. So might
as well decide and modify the JSON object after the decision has been
made whether the quizId stays as JSON or becomes part of the request
URL.