Being able to use predefined pub quizzes has been implemented.
The user can also edit the quiz to his own liking e.g. add and remove
questions.

Two new fields have been added in Quiz:
isPublished - tells whether the quiz has been published for others or
not.
isDuplicate - indicates whether it's a quiz that was generated from
another quiz, these quizzes can not be published.

Currently it creates a name from the name of the duplicated quiz + the
name of the user that created it.