Check if quiz is active within time period. Form validation.
Some refining done to the registration of predefined quiz.

Check for when you're trying to register a predefined quiz whether there
is a quiz with the given room name active within the given time period.
(Room name used from predefined quizzes is now up to the user to
define).

Form validation when trying to register a predefined quiz has been
implemented.